+++ src/CeLux/backends/Decoder.cpp
@@ -2,6 +2,7 @@
 #include "Decoder.hpp"
 #include <Factory.hpp>
 
+
 using namespace celux::error;
 
 namespace celux
@@ -9,7 +10,7 @@ namespace celux
 
 Decoder::Decoder(int numThreads, std::vector<std::shared_ptr<FilterBase>> filters)
     : converter(nullptr), formatCtx(nullptr), codecCtx(nullptr), pkt(nullptr),
-      videoStreamIndex(-1), numThreads(numThreads), filters_(filters)
+      videoStreamIndex(-1), numThreads(numThreads), filters_(filters), remuxer_("")
 {
     CELUX_DEBUG("BASE DECODER: Decoder constructed");
 }
@@ -24,7 +25,7 @@ Decoder::Decoder(Decoder&& other) noexcept
     : formatCtx(std::move(other.formatCtx)), codecCtx(std::move(other.codecCtx)),
       pkt(std::move(other.pkt)), videoStreamIndex(other.videoStreamIndex),
       properties(std::move(other.properties)), frame(std::move(other.frame)),
-      converter(std::move(other.converter))
+      converter(std::move(other.converter)), remuxer_(std::move(other.remuxer_))
 {
     CELUX_DEBUG("BASE DECODER: Decoder move constructor called");
     other.videoStreamIndex = -1;
@@ -253,7 +254,6 @@ bool Decoder::initFilterGraph()
 
     return true;
 }
-
 
 void Decoder::openFile(const std::string& filePath)
 {